Add null checks and move Asset and Tag
