+++ src/main/java/seedu/address/logic/commands/AddCommand.java
@@ -81,6 +81,7 @@ public class AddCommand extends Command {
      * @throws IllegalArgumentException if the user input does not conform the expected format
      */
     public static AddCommand of(String args) throws IllegalArgumentException {
+        requireNonNull(args);
         ArgumentMultimap argMultimap =
                 ArgumentTokenizer.tokenize(args, PREFIX_NAME, PREFIX_PHONE, PREFIX_EMAIL,
                         PREFIX_ADDRESS, PREFIX_TAG, PREFIX_ASSET);
