+++ src/main/java/seedu/address/logic/commands/UndoCommand.java
@@ -1,5 +1,7 @@
 package seedu.address.logic.commands;
 
+import static java.util.Objects.requireNonNull;
+
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 
@@ -18,6 +20,7 @@ public class UndoCommand extends Command {
 
     @Override
     public String execute(Model model) throws CommandException {
+        requireNonNull(model);
         if (!model.canUndo()) {
             throw new CommandException(MESSAGE_UNDO_EXCEPTION);
         }
