+++ src/main/java/seedu/address/model/Model.java
@@ -5,9 +5,9 @@ import java.util.function.Predicate;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
-import seedu.address.model.asset.Asset;
 import seedu.address.model.exceptions.AddressBookException;
 import seedu.address.model.person.Person;
+import seedu.address.model.person.fields.Asset;
 
 /**
  * The API of the Model component.
