+++ src/main/java/seedu/address/model/person/fields/Tags.java
@@ -9,8 +9,6 @@ import java.util.stream.Stream;
 
 import com.fasterxml.jackson.annotation.JsonValue;
 
-import seedu.address.model.tag.Tag;
-
 //@@author aureliony
 /**
  * Represents an abstraction for a list of tags.
