+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -59,8 +59,8 @@ import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyUserPrefs;
-import seedu.address.model.asset.Asset;
 import seedu.address.model.person.Person;
+import seedu.address.model.person.fields.Asset;
 import seedu.address.testutil.PersonBuilder;
 
 public class AddCommandTest {
